fixes and improvements

+++ lib/user_settings/delete_all_data.dart
@@ -6,6 +6,7 @@ import 'package:csocsort_szamla/essentials/widgets/future_success_dialog.dart';
 import 'package:flutter/material.dart';
 import 'package:easy_localization/easy_localization.dart';
 import 'package:provider/provider.dart';
+import 'package:shared_preferences/shared_preferences.dart';
 
 import '../essentials/widgets/gradient_button.dart';
 
@@ -19,15 +20,12 @@ class _DeleteAllDataState extends State<DeleteAllData> {
     try {
       await Http.delete(uri: '/user');
       await clearAllCache();
-      AppStateProvider userProvider = context.read<AppStateProvider>();
-      userProvider.setGroup(null);
-      userProvider.setGroups([]);
-      userProvider.setUser(null);
       return BoolFutureOutput.True;
     } catch (_) {
       throw _;
     }
   }
+
   @override
   Widget build(BuildContext context) {
     return Card(
@@ -39,8 +37,10 @@ class _DeleteAllDataState extends State<DeleteAllData> {
             Center(
                 child: Text(
               'delete_all_data'.tr(),
-              style: Theme.of(context).textTheme.titleLarge!.copyWith(
-                  color: Theme.of(context).colorScheme.onSurfaceVariant),
+              style: Theme.of(context)
+                  .textTheme
+                  .titleLarge!
+                  .copyWith(color: Theme.of(context).colorScheme.onSurfaceVariant),
               textAlign: TextAlign.center,
             )),
             SizedBox(
@@ -49,8 +49,10 @@ class _DeleteAllDataState extends State<DeleteAllData> {
             Center(
               child: Text(
                 'delete_all_data_explanation'.tr(),
-                style: Theme.of(context).textTheme.titleSmall!.copyWith(
-                    color: Theme.of(context).colorScheme.onSurfaceVariant),
+                style: Theme.of(context)
+                    .textTheme
+                    .titleSmall!
+                    .copyWith(color: Theme.of(context).colorScheme.onSurfaceVariant),
                 textAlign: TextAlign.center,
               ),
             ),
@@ -64,25 +66,34 @@ class _DeleteAllDataState extends State<DeleteAllData> {
                   child: Icon(Icons.delete_forever),
                   onPressed: () {
                     showDialog(
-                            builder: (context) => ConfirmChoiceDialog(
-                                  choice: 'sure_user_delete',
-                                ),
-                            context: context)
-                        .then((value) {
-                      if (value ?? false) {
-                        showFutureOutputDialog(
-                          context: context,
-                          future: _deleteAllData(),
-                          outputCallbacks: {
-                            BoolFutureOutput.True: () =>
-                                Navigator.of(context).pushAndRemoveUntil(
+                      builder: (context) => ConfirmChoiceDialog(
+                        choice: 'sure_user_delete',
+                      ),
+                      context: context,
+                    ).then(
+                      (value) {
+                        if (value ?? false) {
+                          showFutureOutputDialog(
+                            context: context,
+                            future: _deleteAllData(),
+                            outputCallbacks: {
+                              BoolFutureOutput.True: () async {
+                                await Navigator.of(context).pushAndRemoveUntil(
                                   MaterialPageRoute(builder: (context) => LoginOrRegisterPage()),
                                   (r) => false,
-                                ),
-                          },
-                        );
-                      }
-                    });
+                                );
+                                AppStateProvider userProvider = context.read<AppStateProvider>();
+                                userProvider.setGroup(null);
+                                userProvider.setGroups([]);
+                                userProvider.setUser(null);
+                                var prefs = await SharedPreferences.getInstance();
+                                await prefs.remove('current_username');
+                              },
+                            },
+                          );
+                        }
+                      },
+                    );
                   },
                 ),
               ],
